creating the navbar and fixing other things in the API

+++ server/controllers/todo.controller.js
@@ -2,10 +2,15 @@ import Todo from "../models/todo.model.js";
 
 export const getTodos = async (req, res, next) => {
   try {
-    const todos = await Todo.find().sort({ createdAt: -1 });
+    let todos;
+    if (req.user.isAdmin) {
+      todos = await Todo.find().populate("user", "name email").sort({ createdAt: -1 });
+    } else {
+      todos = await Todo.find({ user: req.user._id }).sort({ createdAt: -1 });
+    }
     res.json(todos);
   } catch (error) {
-    console.error(err);
+    next(err);
     res.status(500).json({ message: "Server Error" });
   }
 };
@@ -14,12 +19,13 @@ export const addTodo = async (req, res, next) => {
   try {
     const newTodo = new Todo({
       text: req.body.text,
+      user: req.user._id,
     });
 
     const todo = await newTodo.save();
     res.json(todo);
   } catch (err) {
-    console.error(err);
+    next(err);
     res.status(500).json({ message: "Server Error" });
   }
 };
@@ -30,6 +36,12 @@ export const updateTodo = async (req, res, next) => {
 
     if (!todo) {
       return res.status(404).json({ message: "Todo not found" });
+    }
+
+    if (!req.user.isAdmin && todo.user.toString() !== req.user._id.toString()) {
+      return res
+        .status(403)
+        .json({ message: "Unathorized can't update this todo" });
     }
 
     if (req.body.text !== undefined) {
@@ -43,7 +55,7 @@ export const updateTodo = async (req, res, next) => {
     await todo.save();
     res.json(todo);
   } catch (err) {
-    console.error(err);
+    next(err);
     res.status(500).json({ message: "Server Error" });
   }
 };
@@ -56,10 +68,16 @@ export const deleteTodo = async (req, res, next) => {
       return res.status(404).json({ message: "Todo not found" });
     }
 
+    if (!req.user.isAdmin && todo.user.toString() !== req.user._id.toString()) {
+      return res
+        .status(403)
+        .json({ message: "Unauthorized: Cannot delete this todo" });
+    }
+
     await todo.deleteOne();
     res.json({ message: "Todo removed" });
   } catch (err) {
-    console.error(err);
+    next(err);
     res.status(500).json({ message: "Server Error" });
   }
 };
